break apart consensus package by function
